requirements.txt
streamlit
pytesseract
opencv-python
Pillow
openai
